Time: 0 ms (100%), Space: 8.3 MB (74.44%) - LeetHub

+++ LeetCode/Easy/2469-convert-the-temperature/2469-convert-the-temperature.cpp
@@ -7,7 +7,8 @@ public:
         double kelvin = (celsius + 273.15);
         double farenheit = (celsius * 1.80 + 32.00);
 
-        vector<double> ans = {(double)kelvin, (double)farenheit};
+        // vector<double> ans = {(double)kelvin, (double)farenheit};
+        vector<double> ans = {kelvin,farenheit};
         return ans;
     }
 };
